+++ lib/has_global_session/integrated_session.rb
@@ -1,51 +1,53 @@
 module HasGlobalSession
   class IntegratedSession
-    def initialize(local_session, global_session)
-      @local_session = local_session
-      @global_session = global_session
+    attr_reader :local, :global
+    
+    def initialize(local, global)
+      @local = local
+      @global = global
     end
 
     def [](key)
       key = key.to_s
-      if @global_session.supports_key?(key)
-        @global_session[key]
+      if @global.supports_key?(key)
+        @global[key]
       else
-        @local_session[key]
+        @local[key]
       end
     end
 
     def []=(key, value)
       key = key.to_s
-      if @global_session.supports_key?(key)
-        @global_session[key] = value
+      if @global.supports_key?(key)
+        @global[key] = value
       else
-        @local_session[key] = value
+        @local[key] = value
       end
     end
 
     def has_key?(key)
       key = key.to_s
-      @global_session.has_key(key) || @local_session.has_key?(key)
+      @global.has_key(key) || @local.has_key?(key)
     end
 
     def keys
-      @global_session.keys + @local_session.keys
+      @global.keys + @local.keys
     end
 
     def values
-      @global_session.values + @local_session.values
+      @global.values + @local.values
     end
 
     def each_pair(&block)
-      @global_session.each_pair(&block)
-      @local_session.each_pair(&block)
+      @global.each_pair(&block)
+      @local.each_pair(&block)
     end
 
     def method_missing(meth, *args)
-      if @global_session.respond_to?(meth)
-        return @global_session.send(meth, *args)
-      elsif @local_session.respond_to?(meth)
-        return @local_session.send(meth, *args)
+      if @global.respond_to?(meth)
+        return @global.send(meth, *args)
+      elsif @local.respond_to?(meth)
+        return @local.send(meth, *args)
       else
         super
       end